Do not delete repos

+++ vm.tf
@@ -83,24 +83,26 @@ resource "null_resource" "ansible" {
   # Run Ansible Playbook
   provisioner "local-exec" {
     command = <<EOT
-    rm -rf ${var.config_path}/${var.module_name}_ansible
+    if [[ -f "${var.config_path}/${var.module_name}_ansible" ]]; then
+      rm -rf ${var.config_path}/${var.module_name}_ansible
+    fi
     git clone --depth 1 ${var.dots_ansible_repo} ${var.config_path}/${var.module_name}_ansible
     # Remove .git folder
-    rm -rf ${var.config_path}/${var.module_name}_ansible/.git
+    # rm -rf ${var.config_path}/${var.module_name}_ansible/.git
     cd ${var.config_path} && ansible-playbook -i ${var.module_name}_hosts -e "@${var.config_path}/${var.module_name}_host_vars.yml" ${var.module_name}_ansible/${var.module_name}.yml
     EOT
   }
 }
 
-resource "null_resource" "ansible_cleanup" {
-  depends_on = [null_resource.ansible]
-  # Run Ansible Playbook
-  provisioner "local-exec" {
-    command = <<EOT
-    rm -rf ${var.config_path}/${var.module_name}_ansible
-    EOT
-  }
-}
+# resource "null_resource" "ansible_cleanup" {
+#   depends_on = [null_resource.ansible]
+#   # Run Ansible Playbook
+#   provisioner "local-exec" {
+#     command = <<EOT
+#     rm -rf ${var.config_path}/${var.module_name}_ansible
+#     EOT
+#   }
+# }
 
 # Created a resource to clone the repo because configuration
 # files are copied to the repo
@@ -109,10 +111,12 @@ resource "null_resource" "git_clone_hms" {
   # Clone Dots Ansible Project
   provisioner "local-exec" {
     command = <<EOT
-    rm -rf ${var.config_path}/ansible-hms-docker 
+    if [[ -f "${var.config_path}/ansible-hms-docker" ]]; then
+      rm -rf ${var.config_path}/ansible-hms-docker 
+    fi
     git clone --depth 1 https://github.com/prashantsolanki3/ansible-hms-docker ${var.config_path}/ansible-hms-docker
     # Remove .git folder
-    rm -rf ${var.config_path}/ansible-hms-docker/.git
+    # rm -rf ${var.config_path}/ansible-hms-docker/.git
     EOT
   }
 }
